Mudança nos scripts
Mudança em nomes, lógica e código

+++ Assets/Scripts/Database Models/PlayerModel.cs
@@ -7,7 +7,7 @@ public class PlayerModel
     public string Name { get; set; }
     
     [Column("points"), NotNull]
-    public int Points { get; set; } = 0;
+    public float Points { get; set; } = 0;
 
     [Column("time"), NotNull]
     public float Time { get; set; } = 0;
